changed most of IntegerFields to BigIntegerFields in accounting models

+++ accounting/models.py
@@ -18,7 +18,7 @@ class CostType(TimestampedModel):
 
 class Cost(TimestampedModel):
     type = models.ForeignKey(CostType, on_delete=models.PROTECT, related_name='costs')
-    amount = models.IntegerField()
+    amount = models.BigIntegerField()
     date = models.DateField()
     description = models.TextField(default='', blank=True)
 
@@ -27,7 +27,7 @@ class Cost(TimestampedModel):
 
 
 class Income(TimestampedModel):
-    amount = models.IntegerField()
+    amount = models.BigIntegerField()
     date = models.DateField()
     description = models.TextField(default='', blank=True)
 
@@ -36,7 +36,7 @@ class Income(TimestampedModel):
 
 
 class ProductCost(TimestampedModel):
-    amount = models.IntegerField()
+    amount = models.BigIntegerField()
     date = models.DateField()
     description = models.TextField(default='', blank=True)
 
@@ -45,7 +45,7 @@ class ProductCost(TimestampedModel):
 
 
 class Invoice(models.Model):
-    number = models.IntegerField(unique=True)
+    number = models.BigIntegerField(unique=True)
     start_date_persian = models.CharField(max_length=255)
     end_date_persian = models.CharField(max_length=255)
     start_date = models.DateField(blank=True, null=True)
@@ -60,21 +60,21 @@ class InvoiceItem(models.Model):
     )
 
     row_number = models.IntegerField()
-    code = models.IntegerField()
+    code = models.BigIntegerField()
 
     date_persian = models.CharField(max_length=255)
     date = models.DateTimeField(blank=True, null=True)
 
-    dkpc = models.IntegerField()
+    dkpc = models.BigIntegerField()
     variant_title = models.CharField(max_length=255)
-    order_id = models.IntegerField()
+    order_id = models.BigIntegerField()
     serial = models.CharField(max_length=255)
-    credit = models.IntegerField()
-    debit = models.IntegerField()
-    credit_discount = models.IntegerField()
-    debit_discount = models.IntegerField()
-    credit_final = models.IntegerField()
-    debit_final = models.IntegerField()
+    credit = models.BigIntegerField()
+    debit = models.BigIntegerField()
+    credit_discount = models.BigIntegerField()
+    debit_discount = models.BigIntegerField()
+    credit_final = models.BigIntegerField()
+    debit_final = models.BigIntegerField()
     description = models.TextField(blank=True, null=True)
     calculated = models.BooleanField(default=False)
